Fixed workflow problems and played
I was having a couple of problems with the workflow
as it was originally setup. Firstly, Grunt would
occasionally get stuck in an endless loop of updating
files. I think I've resolved this by breaking the
watch command up into separate parts.

The other problem I was having was to do with Vagrant.
When I was editing main.js with large changes and
refreshing the page Vagrant was serving up, I saw my
changes. However, with small changes, the effects of
them were not being reflected on the page. Apparently
this is a known issue due to VirtualBox and the
sendfile mechanism. This is sorted by changing "sendfile on"
to "sendfile off" in /etc/nginx/nginx.conf. However,
I couldn't think of an elegant way of changing that
so I just copied the entire file into bootstrap.sh
and got it to echo the entire file with the change to
/etc/nginx/nginx.conf.

Finally, I included some mouse input into the main.js
example.