Implemented Hotline Miami-esque Controls
The commit title says it all really. The player controlled
character (called 'protagonist' as I couldn't be arsed to
draw anything new) looks towards the mouse and walks around
using WASD. The other thing to note is that I've tried to
oraganise the project nicely, with the player 'class' being
a separate file.